C-B30 | Primera gran parte de la refactorización completada
