+++ components/footer/footer.html
@@ -6,7 +6,7 @@
       <p class="col-md-4 mb-0 text-body-secondary">© 2024 Ayotl</p>
 
       <a
-        data-link="inicio"
+        data-link="inicio noModule"
         class="col-md-4 d-flex align-items-center justify-content-center mb-3 mb-md-0 me-md-auto link-body-emphasis text-decoration-none"
       >
         <img class="logo-size" src="../../assets/imgs/logo.svg" alt="" />
