+++ components/footer/footer.js
@@ -1,19 +1,12 @@
+const footer = document.querySelector('footer');
 
-export default async function traerFooter(URL) {
-   try {
-      const respuesta = await fetch(URL);
-      // Verificar si la respuesta fue exitosa
-      if (!respuesta.ok) {
-         throw new Error(`Error en la solicitud: ${response.status}`);
-      }
-      
-      const htmlFooter = await respuesta.text();
-      // console.log (htmlFooter);
-      
-      return htmlFooter;
-      
-   } catch (error) {
-      console.error("Error al obtener la footer:", error);
-   }
-}
+footer.addEventListener('click', (event) => {
+  if (event.target.tagName === 'A') { // Verifica si el target es un enlace
+    event.preventDefault(); // Evita el comportamiento predeterminado
 
+    // Ejecuta tu código aquí
+    console.log(`Enlace clickeado: ${event.target.href}`);
+    
+    // Lógica adicional si es necesario
+  }
+});
